[code refactoring] renamed components to more meaningful names
