+++ src/components/ActivtyModal.js
@@ -8,9 +8,104 @@ import DialogContentText from "@material-ui/core/DialogContentText";
 import DialogTitle from "@material-ui/core/DialogTitle";
 import TextField from "@material-ui/core/TextField";
 import moment from "moment";
-import Calender from "./calender/Calender";
+import Calendar from "./calendar/Calendar";
 import Typography from "@material-ui/core/Typography";
 import { pink, green } from "@material-ui/core/colors";
+
+export default function ActivtyModal(props) {
+  const classes = useStyles();
+  const [viewAll, setViewAll] = useState(false);
+  let myEvents = [];
+
+  const formatDate = (dateTime, title) => {
+    const date = dateTime.substring(0, dateTime.lastIndexOf(" ") + 1);
+    const time =
+      " " + dateTime.substring(dateTime.lastIndexOf(" ") + 1, dateTime.length);
+
+    const formattedDate = moment(date).format("YYYY-MM-DD");
+    myEvents = [...myEvents, { title: title + time, date: formattedDate }];
+
+    return formattedDate + time;
+  };
+
+  useEffect(() => {
+    setViewAll(false);
+  }, [props]);
+
+  const renderDate = () => {
+    return props.data.activity_periods.map((item, index) => {
+      const startDate = formatDate(item.start_time, "start");
+      const endDate = formatDate(item.end_time, "end");
+      return (
+        <Fragment key={index}>
+          <TextField
+            id="datetime-start"
+            label="Start Time"
+            type="text"
+            value={startDate}
+            className={classes.textField}
+            InputLabelProps={{
+              shrink: true
+            }}
+            disabled
+          />
+          <TextField
+            id="datetime-end"
+            label="End Time"
+            type="text"
+            value={endDate}
+            className={classes.textField}
+            InputLabelProps={{
+              shrink: true
+            }}
+            disabled
+          />
+        </Fragment>
+      );
+    });
+  };
+
+  return (
+    <Dialog
+      maxWidth="md"
+      open={props.open}
+      onClose={props.hanndleClose}
+      aria-labelledby="max-width-dialog-title"
+    >
+      <DialogTitle id="max-width-dialog-title">
+        <Typography className={classes.green}>
+          {"User Name: "}
+          <span className={classes.pink}>{props.data.real_name}</span>
+        </Typography>
+      </DialogTitle>
+      <DialogContent>
+        <DialogContentText>Your Activity Period</DialogContentText>
+        {props.data.activity_periods && renderDate()}
+        {viewAll && (
+          <div className={classes.calender}>
+            <Calendar eventsList={myEvents} />
+          </div>
+        )}
+      </DialogContent>
+      <DialogActions className={classes.actionButtons}>
+        <Button
+          onClick={() => setViewAll(!viewAll)}
+          variant="outlined"
+          color="primary"
+        >
+          {viewAll ? "Hide Calendar" : "View All"}
+        </Button>
+        <Button
+          onClick={props.hanndleClose}
+          variant="outlined"
+          color="secondary"
+        >
+          Close
+        </Button>
+      </DialogActions>
+    </Dialog>
+  );
+}
 
 const useStyles = makeStyles(theme => ({
   form: {
@@ -49,94 +144,3 @@ const useStyles = makeStyles(theme => ({
     marginTop: "2rem"
   }
 }));
-
-export default function MaxWidthDialog(props) {
-  const classes = useStyles();
-  const [viewAll, setViewAll] = useState(false);
-  let myEvents = [];
-  const formatDate = (dateTime, title) => {
-    const date = dateTime.substring(0, dateTime.lastIndexOf(" ") + 1);
-    const time =
-      " " + dateTime.substring(dateTime.lastIndexOf(" ") + 1, dateTime.length);
-    const formattedDate = moment(date).format("YYYY-MM-DD");
-    myEvents = [...myEvents, { title: title + time, date: formattedDate }];
-    return formattedDate + time;
-  };
-  useEffect(() => {
-    setViewAll(false);
-  }, [props]);
-  const renderDate = () => {
-    return props.data.activity_periods.map((item, index) => {
-      const startDate = formatDate(item.start_time, "start");
-      const endDate = formatDate(item.end_time, "end");
-      return (
-        <Fragment key={index}>
-          <TextField
-            id="datetime-start"
-            label="Start Time"
-            type="text"
-            value={startDate}
-            className={classes.textField}
-            InputLabelProps={{
-              shrink: true
-            }}
-            disabled
-          />
-          <TextField
-            id="datetime-end"
-            label="End Time"
-            type="text"
-            value={endDate}
-            className={classes.textField}
-            InputLabelProps={{
-              shrink: true
-            }}
-            disabled
-          />
-        </Fragment>
-      );
-    });
-  };
-  return (
-    <React.Fragment>
-      <Dialog
-        maxWidth="md"
-        open={props.open}
-        onClose={props.hanndleClose}
-        aria-labelledby="max-width-dialog-title"
-      >
-        <DialogTitle id="max-width-dialog-title">
-          <Typography className={classes.green}>
-            {"User Name: "}
-            <span className={classes.pink}>{props.data.real_name}</span>
-          </Typography>
-        </DialogTitle>
-        <DialogContent>
-          <DialogContentText>Your Activity Period</DialogContentText>
-          {props.data.activity_periods && renderDate()}
-          {viewAll && (
-            <div className={classes.calender}>
-              <Calender eventsList={myEvents} />
-            </div>
-          )}
-        </DialogContent>
-        <DialogActions className={classes.actionButtons}>
-          <Button
-            onClick={() => setViewAll(!viewAll)}
-            variant="outlined"
-            color="primary"
-          >
-            {viewAll ? "Hide Calender" : "View All"}
-          </Button>
-          <Button
-            onClick={props.hanndleClose}
-            variant="outlined"
-            color="secondary"
-          >
-            Close
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </React.Fragment>
-  );
-}
